[#171] api 파일 정리
