+++ src/app/dashboard/[id]/components/ColumnCreateForm.tsx
@@ -1,6 +1,6 @@
 import { useForm } from 'react-hook-form'
 
-import api from '@/app/utils/axiosInstance'
+import api from '@/lib/axiosInstance'
 import Button from '@/components/Button'
 import ConfirmModalContent from '@/components/ConfirmModalContent'
 import Form from '@/components/Form'
