+++ src/app/dashboard/[id]/components/ColumnEditForm.tsx
@@ -2,7 +2,7 @@ import axios from 'axios'
 import { useForm } from 'react-hook-form'
 
 import Close from '/public/icons/close.svg'
-import api from '@/app/utils/axiosInstance'
+import api from '@/lib/axiosInstance'
 import Button from '@/components/Button'
 import ConfirmModalContent from '@/components/ConfirmModalContent'
 import DeleteAlertModal from '@/components/DeleteAlertModal'
