+++ src/app/dashboard/[id]/components/TaskCardContent.tsx
@@ -5,7 +5,7 @@ import Close from '/public/icons/close.svg'
 import Kebab from '/public/icons/kebab-menu.svg'
 import ColumnChip from '@/app/dashboard/[id]/components/ColumnChip'
 import Comment from '@/app/dashboard/[id]/components/Comment'
-import api from '@/app/utils/axiosInstance'
+import api from '@/lib/axiosInstance'
 import Button from '@/components/Button'
 import Chip from '@/components/Chip'
 import ConfirmModalContent from '@/components/ConfirmModalContent'
