+++ src/app/login/components/LoginForm.tsx
@@ -4,8 +4,8 @@ import axios from 'axios'
 import { useRouter } from 'next/navigation'
 import { useForm } from 'react-hook-form'
 
-import api from '@/app/utils/axiosInstance'
-import { getDashboardList } from '@/app/utils/dashboardsApi'
+import api from '@/lib/axiosInstance'
+import { getDashboardList } from '@/lib/dashboardsApi'
 import Button from '@/components/Button'
 import ConfirmModalContent from '@/components/ConfirmModalContent'
 import Form from '@/components/Form'
