+++ src/app/mypage/components/PasswordEditForm.tsx
@@ -3,7 +3,7 @@
 import axios from 'axios'
 import { useForm } from 'react-hook-form'
 
-import api from '@/app/utils/axiosInstance'
+import api from '@/lib/axiosInstance'
 import Button from '@/components/Button'
 import ConfirmModalContent from '@/components/ConfirmModalContent'
 import Form from '@/components/Form'
