+++ src/app/mypage/components/ProfileEditForm.tsx
@@ -6,8 +6,8 @@ import { useForm } from 'react-hook-form'
 
 import Edit from '/public/icons/edit.svg'
 import Plus from '/public/icons/plus.svg'
-import { imageUpload } from '@/app/utils/api'
-import api from '@/app/utils/axiosInstance'
+import { imageUpload } from '@/lib/api'
+import api from '@/lib/axiosInstance'
 import Button from '@/components/Button'
 import ConfirmModalContent from '@/components/ConfirmModalContent'
 import Form from '@/components/Form'
