+++ src/components/AddTaskModal.tsx
@@ -10,8 +10,8 @@ import AddBox from '/public/icons/add-box2.svg'
 import Calendar from '/public/icons/calendar.svg'
 import Edit from '/public/icons/edit.svg'
 import AsigneeSelect from '@/app/dashboard/[id]/components/AsigneeSelect'
-import { imageUpload } from '@/app/utils/api'
-import api from '@/app/utils/axiosInstance'
+import { imageUpload } from '@/lib/api'
+import api from '@/lib/axiosInstance'
 import Form from '@/components/Form'
 import useModalStore from '@/store/useModalStore'
 import useUserStore from '@/store/useUserStore'
