+++ src/components/DashboardFeature.tsx
@@ -8,8 +8,8 @@ import { useForm } from 'react-hook-form'
 import IconAddBox from '/public/icons/add-box.svg'
 import IconClose from '/public/icons/close.svg'
 import IconSettings from '/public/icons/settings.svg'
-import api from '@/app/utils/axiosInstance'
-import { deleteDashboard } from '@/app/utils/dashboardsApi'
+import api from '@/lib/axiosInstance'
+import { deleteDashboard } from '@/lib/dashboardsApi'
 import useDashboardStore from '@/store/useDashboardStore'
 import useModalStore from '@/store/useModalStore'
 import { Dashboard } from '@/types/types'
