+++ src/layouts/DashboardCol.tsx
@@ -5,7 +5,7 @@ import { useEffect, useState } from 'react'
 import Bullet from '/public/icons/bullet.svg'
 import Setting from '/public/icons/settings.svg'
 import ColumnEditForm from '@/app/dashboard/[id]/components/ColumnEditForm'
-import api from '@/app/utils/axiosInstance'
+import api from '@/lib/axiosInstance'
 import AddTaskModal from '@/components/AddTaskModal'
 import NewTaskButton from '@/components/NewTaskButton'
 import TaskCard from '@/components/TaskCard'
